+++ CHANGELOG.md
@@ -5,10 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.5.0] (Unreleased)
+## [0.5.0] (2026-01-30)
 - Added AnyPcapReader to read both pcap and pcapng files
 - Introduce AsyncPcapNgReader for asynchronous pcapng reading
 - Added AsyncAnyPcapReader for asynchronous reading of both pcap and pcapng files
+- Cargo Check Fix
 
 ## [0.4.0] (2026-01-06)
 - Fixed Bad Error Name